Mv descrstats, sampleScatter, singleCellPlots, & singlecellSamples route code to server/src/routes
