+++ server/routes/termdb.violinBox.ts
@@ -16,7 +16,7 @@ import { run_R } from '@sjcrh/proteinpaint-r'
 import { getData } from '../src/termdb.matrix.js'
 import { createCanvas } from 'canvas'
 import { getOrderedLabels } from '../src/termdb.barchart.js'
-import { getDescrStats, getStdDev, getMean } from './termdb.descrstats.ts'
+import { getDescrStats, getStdDev, getMean } from '#routes/termdb.descrstats.ts'
 import { isNumericTerm } from '#shared/terms.js'
 import { boxplot_getvalue } from '../src/utils.js'
 import { roundValueAuto } from '#shared/roundValue.js'
