+++ server/src/mds3.init.js
@@ -35,11 +35,11 @@ import {
 import { get_samples, get_active_groupset } from './termdb.sql.js'
 import { server_init_db_queries } from './termdb.server.init.ts'
 import { barchart_data } from './termdb.barchart.js'
-import { mayInitiateScatterplots } from '../routes/termdb.sampleScatter.ts'
+import { mayInitiateScatterplots } from '#routes/termdb.sampleScatter.ts'
 import { mayInitiateMatrixplots, mayInitiateNumericDictionaryTermplots } from './termdb.matrix.js'
 import { add_bcf_variant_filter } from './termdb.snp.js'
 import { validate_correlationVolcano } from './routes/correlationVolcano.ts'
-import { validate_query_singleCell } from '../routes/termdb.singlecellSamples.ts'
+import { validate_query_singleCell } from '#routes/termdb.singlecellSamples.ts'
 import { validate_query_proteome } from '../routes/termdb.proteome.ts'
 import { validate_query_TopVariablyExpressedGenes } from '#routes/termdb.topVariablyExpressedGenes.ts'
 import { validate_query_singleSampleMutation } from '#routes/termdb.singleSampleMutation.ts'
@@ -160,62 +160,62 @@ export async function init(ds, genome, totalDsLst = 0) {
 		if (response?.status != 'OK') throw response
 	}
 
-  try {
-	// must validate termdb first
-	await validate_termdb(ds)
-	validateDemoJwtInputs(ds)
+	try {
+		// must validate termdb first
+		await validate_termdb(ds)
+		validateDemoJwtInputs(ds)
 
-	if (ds.queries) {
-		// must validate snvindel query before variant2sample
-		// as vcf header must be parsed to supply samples for variant2samples
-		await validate_query_snvindel(ds, genome)
-		await validate_query_svfusion(ds, genome)
-		await validate_query_geneCnv(ds, genome)
-		await validate_query_cnv(ds, genome)
-		await validate_query_itd(ds, genome)
-		await validate_query_ld(ds, genome)
-		await validate_query_geneExpression(ds, genome)
-		await validateQueryIsoformExpression(ds, genome)
-		await validate_query_ssGSEA(ds, genome)
-		await validate_query_dnaMethylation(ds, genome)
-		await validate_query_metaboliteIntensity(ds, genome)
-		await validate_query_proteome(ds, genome)
-		await validate_query_getTopTermsByType(ds, genome)
-		await validate_query_getTopMutatedGenes(ds, genome)
-		await validate_query_getSampleImages(ds, genome)
-		await validate_query_getWSIAnnotations(ds)
-		await validate_query_getWSIClassesQuery(ds)
-		await validate_query_getSampleWSImages(ds, genome)
-		await validate_query_deleteWSIAnnotation(ds)
-		await validate_query_saveWSIAnnotation(ds)
-		await validate_query_getWSISamples(ds, genome)
-		await makeAdHocDicTermdbQueries(ds)
-		await validate_query_rnaseqGeneCount(ds, genome)
-		await validate_query_singleSampleMutation(ds, genome)
-		await validate_query_singleSampleGenomeQuantification(ds, genome)
-		await validate_query_singleSampleGbtk(ds, genome)
-		//await validate_query_probe2cnv(ds, genome)
-		await validate_query_singleCell(ds, genome)
-		await validate_query_TopVariablyExpressedGenes(ds)
-		await validate_query_trackLst(ds, genome)
+		if (ds.queries) {
+			// must validate snvindel query before variant2sample
+			// as vcf header must be parsed to supply samples for variant2samples
+			await validate_query_snvindel(ds, genome)
+			await validate_query_svfusion(ds, genome)
+			await validate_query_geneCnv(ds, genome)
+			await validate_query_cnv(ds, genome)
+			await validate_query_itd(ds, genome)
+			await validate_query_ld(ds, genome)
+			await validate_query_geneExpression(ds, genome)
+			await validateQueryIsoformExpression(ds, genome)
+			await validate_query_ssGSEA(ds, genome)
+			await validate_query_dnaMethylation(ds, genome)
+			await validate_query_metaboliteIntensity(ds, genome)
+			await validate_query_proteome(ds, genome)
+			await validate_query_getTopTermsByType(ds, genome)
+			await validate_query_getTopMutatedGenes(ds, genome)
+			await validate_query_getSampleImages(ds, genome)
+			await validate_query_getWSIAnnotations(ds)
+			await validate_query_getWSIClassesQuery(ds)
+			await validate_query_getSampleWSImages(ds, genome)
+			await validate_query_deleteWSIAnnotation(ds)
+			await validate_query_saveWSIAnnotation(ds)
+			await validate_query_getWSISamples(ds, genome)
+			await makeAdHocDicTermdbQueries(ds)
+			await validate_query_rnaseqGeneCount(ds, genome)
+			await validate_query_singleSampleMutation(ds, genome)
+			await validate_query_singleSampleGenomeQuantification(ds, genome)
+			await validate_query_singleSampleGbtk(ds, genome)
+			//await validate_query_probe2cnv(ds, genome)
+			await validate_query_singleCell(ds, genome)
+			await validate_query_TopVariablyExpressedGenes(ds)
+			await validate_query_trackLst(ds, genome)
 
-		await validate_variant2samples(ds)
-		await validate_ssm2canonicalisoform(ds)
+			await validate_variant2samples(ds)
+			await validate_ssm2canonicalisoform(ds)
 
-		await mayAdd_refseq2ensembl(ds, genome)
+			await mayAdd_refseq2ensembl(ds, genome)
 
-		await mayAdd_mayGetGeneVariantData(ds, genome)
-	}
+			await mayAdd_mayGetGeneVariantData(ds, genome)
+		}
 
-	await mayValidateAssayAvailability(ds)
-	await mayValidateViewModes(ds)
+		await mayValidateAssayAvailability(ds)
+		await mayValidateViewModes(ds)
 
-	// uncomment below to manually trigger server crash if there is only 1 dataset;
-	// make sure that serverconfig only has one genome and datasets[] entry,
-	// and that the ds.label below matches that entry
-	// if (ds.label == 'GDC') {ds.init = {status: 'fatalError', fatalError: 'test server crash'}; throw ds.init.fatalError}
+		// uncomment below to manually trigger server crash if there is only 1 dataset;
+		// make sure that serverconfig only has one genome and datasets[] entry,
+		// and that the ds.label below matches that entry
+		// if (ds.label == 'GDC') {ds.init = {status: 'fatalError', fatalError: 'test server crash'}; throw ds.init.fatalError}
 
-	if (ds.cohort?.db?.refresh) throw `!!! ds.cohort.db.refresh has been deprecated !!!`
+		if (ds.cohort?.db?.refresh) throw `!!! ds.cohort.db.refresh has been deprecated !!!`
 	} catch (e) {
 		if (!ds.init) ds.init = {}
 		if (ds.init.step != 'gdcBuildDictionary()' || !ds.init.recoverableError) {
