+++ server/src/termdb.js
@@ -6,7 +6,7 @@ import { get_survival } from './termdb.survival.js'
 import { get_regression } from './termdb.regression.js'
 import { validate as snpValidate } from './termdb.snp.js'
 import { isUsableTerm } from '#shared/termdb.usecase.js'
-import { trigger_getLowessCurve } from '../routes/termdb.sampleScatter.ts'
+import { trigger_getLowessCurve } from '#routes/termdb.sampleScatter.ts'
 import { get_mds3variantData } from './mds3.variant.js'
 import { get_lines_bigfile } from './utils.js'
 import { authApi } from './auth.js'
